feat: separate zk-catalog updates
