+++ packages/database/src/index.ts
@@ -83,3 +83,7 @@ export type Database = ReturnType<typeof createRepositories>
 
 export type { CurrentPrice } from './current-price'
 export type { Stake } from './stake'
+export type {
+  VerifierStatus,
+  VerifierStatusRepository,
+} from './verifier-status'
