+++ packages/database/src/verifier-status/index.ts
@@ -1,0 +1,2 @@
+export type { VerifierStatus } from './entity'
+export { VerifierStatusRepository } from './repository'
